Move translation methods
