+++ tests/test_babel_extractors.py
@@ -21,7 +21,8 @@ schema = b"""{
         }]
     },
     "description": {
-        "title": "  zzz  "
+        "title": "  zzz  ",
+        "description": "    "
     }
 }"""
